populates district db

+++ db/migrate/20151102234654_create_districts.rb
@@ -2,7 +2,7 @@ class CreateDistricts < ActiveRecord::Migration
   def change
     create_table :districts do |t|
       t.integer :number
-      t.integer :attendance
+      t.float :attendance
       t.integer :borough_id
 
       t.timestamps null: false
